S-Bench 2/evaluations/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.28.0

# For MCP protocol support
fastmcp>=0.1.0
mcp>=0.1.0

# For vLLM server (optional - install in separate environment)
# vllm>=0.5.4

# For metrics
numpy>=1.24.0

# Optional for better performance
accelerate>=0.20.0